component directory moved, imports edited
